[Rename] 파일 경로 이동 - chat 컴포넌트 파일 경로 이동
