+++ src/pages/Chat/Chatroom.jsx
@@ -1,9 +1,9 @@
 import ChatHeader from '@/components/Chat/ChatHeader';
-import ChatBoard from './ChatBoard';
+import ChatBoard from '../../components/Chat/ChatBoard';
 import { useParams } from 'react-router-dom';
 
 export function Chatroom() {
-    const { roomId } = useParams();
+  const { roomId } = useParams();
   return (
     <div>
       <ChatHeader title={'EUID 피그마 스터디 하실분~'} people={3} />
